Fix up so stuff imports
