+++ .idea/runConfigurations/modelsTest.xml
@@ -6,6 +6,7 @@
     <envs>
       <env name="PYTHONUNBUFFERED" value="1" />
       <env name="DJANGO_SETTINGS_MODULE" value="mobot.settings" />
+      <env name="SECRET_KEY" value="Test" />
     </envs>
     <option name="SDK_HOME" value="$USER_HOME$/.local/share/virtualenvs/mobot-EWD5ongg/bin/python" />
     <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/mobot/apps/merchant_services/tests" />
