+++ app_scripts/admin_start.sh
@@ -19,6 +19,12 @@ SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 python /app/mobot/manage.py makemigrations exchange
 python /app/mobot/manage.py migrate exchange
 
+python /app/mobot/manage.py makemigrations chat
+python /app/mobot/manage.py migrate chat
+
+python /app/mobot/manage.py makemigrations common
+python /app/mobot/manage.py migrate common
+
 python /app/mobot/manage.py makemigrations payment_service
 python /app/mobot/manage.py migrate payment_service
