+++ mobot/apps/chat/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
-class MobotClientAppsConfig(AppConfig):
+class MobotChatAppsConfig(AppConfig):
     name = 'mobot.apps.chat'
     default_auto_field = 'django.db.models.BigAutoField'
     verbose_name = 'Mobot Chat Bot'
