+++ mobot/apps/drop/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
-class MobotClientAppsConfig(AppConfig):
+class DropAppConfig(AppConfig):
     name = 'mobot.apps.drop'
     default_auto_field = 'django.db.models.BigAutoField'
-    verbose_name = 'Serializers etc'
+    verbose_name = 'Drop config'
