+++ mobot/apps/merchant_services/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from mobot.apps.merchant_services.models import Store, Customer, CustomerStorePreferences, Merchant, Product, Shipment, Order, Campaign, UserAccount
 from mobot.apps.payment_service.models import Payment, Transaction
-from mobot.apps.chat import Message
+from mobot.apps.chat.models import Message
 
 
 @admin.register(Store)
