+++ mobot/settings/common.py
@@ -51,13 +51,14 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'django_celery_results',
     'djmoney',
-    'address',
     'djmoney.contrib.exchange',
     'django_fsm',
+    'address',
     'mobot.apps.common.apps.CommonAppConfig',
+    'mobot.apps.drop.apps.DropAppConfig',
+    'mobot.apps.chat.apps.MobotChatAppsConfig',
     'mobot.apps.payment_service.apps.PaymentServiceAppConfig',
     'mobot.apps.merchant_services.apps.MerchantServicesConfig',
-    'mobot.apps.drop.apps.MobotClientAppsConfig',
     'mobot.apps.mobot_web.apps.MobotWebConfig'
 ]
